refactor: s3.type → presigned-url-info.type으로 파일명 변경
